+++ Package.swift
@@ -26,5 +26,9 @@ let package = Package(
                 .unsafeFlags(["-Xlinker", "-sectcreate", "-Xlinker", "__TEXT", "-Xlinker", "__info_plist", "-Xlinker", "Supporting/Info.plist"])
             ]
         ),
+        .testTarget(
+            name: "QuiperTests",
+            dependencies: ["Quiper"]
+        ),
     ]
 )